requirements.txt
flask==2.3.3
torch==2.0.1
transformers==4.33.2
Pillow==10.0.1
werkzeug==2.3.7 